Refactor remaining tests
