+++ Tests/CouchDBTests/CouchDBTest.swift
@@ -52,17 +52,18 @@ class CouchDBTest: XCTestCase {
 
     var database: Database?
 
-	// MARK: - Initializers and test set-up and tear-down
+    // MARK: - Initializers and test set-up and tear-down
 
-	override func setUp() {
-        dropDatabaseIfExists()
+    override func setUp() {
+        delay(dropDatabaseIfExists)
     }
     
     override func tearDown() {
-        dropDatabaseIfExists()
+        delay(dropDatabaseIfExists)
     }
     
     /// Drop the test database, if it exists.
+    ///
     func dropDatabaseIfExists() {
           // Check if DB exists
           couchDBClient.dbExists(dbName) { exists, error in
@@ -78,7 +79,35 @@ class CouchDBTest: XCTestCase {
 
     /// Create the test database, failing the test if it already exists, or if there
     /// is a connectivity error.
+    ///
     func createDatabase() {
+        delay(delayedCreateDatabase)
+    }
+
+    /// Drop the test database, failing the test if it does not exist, or if there
+    /// is a connectivity error.
+    ///
+    func dropDatabase() {
+        delay(delayedDropDatabase)
+    }
+
+    /// Delay an action by a specified amount of time (default 1 second). The purpose
+    /// of delaying actions in CouchDB tests is to avoid exceeding the API limit for
+    /// the test database provider while running multiple executions via Travis.
+    ///
+    func delay(time: Double = 1.0, _ work: @escaping () -> Void) {
+        let start = DispatchSemaphore(value: 0)
+        let end = DispatchSemaphore(value: 0)
+        DispatchQueue.global().asyncAfter(deadline: .now() + time) {
+            start.wait()
+            work()
+            end.signal()
+        }
+        start.signal()
+        end.wait()
+    }
+
+    fileprivate func delayedCreateDatabase() {
         couchDBClient.createDB(dbName) { database, error in
             if let error = error {
                 XCTFail("DB creation error: \(error.code) \(error.localizedDescription)")
@@ -91,20 +120,18 @@ class CouchDBTest: XCTestCase {
             print("Database \"\(self.dbName)\" successfully created")
             self.database = database
         }
-	}
+    }
 
-    /// Drop the test database, failing the test if it does not exist, or if there
-    /// is a connectivity error.
-    func dropDatabase() {
-		// Retrieve and delete test database
-		couchDBClient.deleteDB(dbName) { error in
-			if let error = error {
-				XCTFail("DB deletion error: \(error.code) \(error.localizedDescription)")
-				return
-			}
-			print("Database \"\(self.dbName)\" successfully deleted")
-		}
-	}
-
+    fileprivate func delayedDropDatabase() {
+        // Retrieve and delete test database
+        couchDBClient.deleteDB(dbName) { error in
+            if let error = error {
+                XCTFail("DB deletion error: \(error.code) \(error.localizedDescription)")
+                return
+            }
+            print("Database \"\(self.dbName)\" successfully deleted")
+        }
+    }
+    
 }
